+++ config/routes.rb
@@ -99,13 +99,13 @@ Rails.application.routes.draw do
 
     resource :user, only: [:show] do
       get 'icn', to: 'users#icn'
-      get 'linked_user_accounts', to: 'users#linked_user_accounts'
-      post 'delegate_access', to: 'users#delegate_access'
       collection do
         get 'credential_emails'
       end
       resource :mhv_user_account, only: [:show], controller: 'user/mhv_user_accounts'
     end
+
+    resource :user_account_delegations, only: %i[create show destroy]
 
     resource :veteran_onboarding, only: %i[show update]
